+++ packages/orchestration/src/orchestration-client.test.ts
@@ -333,7 +333,7 @@ describe('orchestration service client', () => {
         template: [{ role: 'user', content: "What's my name?" }]
       }
     };
-    const prompt = {
+    const prompt: Prompt = {
       messagesHistory: [
         {
           role: 'system',
@@ -373,6 +373,143 @@ describe('orchestration service client', () => {
       prompt
     );
     expect(response.data).toEqual(mockResponse);
+  });
+
+  it('calls chatCompletion with template passed as YAML config', async () => {
+    const yamlTemplate = await parseFileToString(
+      'orchestration',
+      'orchestration-chat-completion-yaml-template.yaml'
+    );
+    const configWithYaml: OrchestrationModuleConfig = {
+      llm: {
+        model_name: 'gpt-4o',
+        model_params: { max_tokens: 500 }
+      },
+      templating: yamlTemplate
+    };
+
+    const config: OrchestrationModuleConfig = {
+      llm: {
+        model_name: 'gpt-4o',
+        model_params: { max_tokens: 500 }
+      },
+      templating: {
+        template: [
+          {
+            role: 'system',
+            content:
+              'You are a world-famous poet who can write virtuosic and brilliant poetry on any topic.'
+          },
+          {
+            role: 'user',
+            content:
+              'Write a 1 verse poem about the following topic: {{?topic}}'
+          }
+        ],
+        response_format: {
+          type: 'json_schema',
+          json_schema: {
+            name: 'poem_structure',
+            description: 'Structured format for the generated poem',
+            strict: true,
+            schema: {
+              type: 'object',
+              additionalProperties: false,
+              required: ['title', 'verses', 'theme'],
+              properties: {
+                title: {
+                  type: 'string',
+                  description: 'The title of the poem'
+                },
+                theme: {
+                  type: 'string',
+                  description: 'The central theme or subject of the poem'
+                },
+                verses: {
+                  type: 'array',
+                  description: 'A list of verses making up the poem',
+                  items: {
+                    type: 'string',
+                    description: 'A single verse of the poem'
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    };
+
+    const mockResponse = await parseMockResponse<CompletionPostResponse>(
+      'orchestration',
+      'orchestration-chat-completion-yaml-template-response.json'
+    );
+    mockInference(
+      {
+        data: constructCompletionPostRequest(config, {
+          inputParams: { topic: 'Generative AI Hub' }
+        })
+      },
+      {
+        data: mockResponse,
+        status: 200
+      },
+      {
+        url: 'inference/deployments/1234/completion'
+      }
+    );
+    const response = await new OrchestrationClient(
+      configWithYaml
+    ).chatCompletion({ inputParams: { topic: 'Generative AI Hub' } });
+    expect(response.data).toEqual(mockResponse);
+  }, 60000);
+
+  it('fails when template is an empty string', async () => {
+    const invalidConfigWithYaml: OrchestrationModuleConfig = {
+      llm: {
+        model_name: 'gpt-4o',
+        model_params: { max_tokens: 500 }
+      },
+      templating: ''
+    };
+
+    expect(() =>
+      new OrchestrationClient(invalidConfigWithYaml).chatCompletion({
+        inputParams: { topic: 'Generative AI Hub' }
+      })
+    ).toThrowErrorMatchingInlineSnapshot(
+      '"Templating YAML string must be non-empty."'
+    );
+  });
+
+  it('fails when template YAML string does not conform to the expected specification', async () => {
+    const invalidConfigWithYaml: OrchestrationModuleConfig = {
+      llm: {
+        model_name: 'gpt-4o',
+        model_params: { max_tokens: 500 }
+      },
+      templating: `
+      name: poem
+      version: 0.0.1
+      scenario: agent-evaluator
+      `
+    };
+
+    expect(() =>
+      new OrchestrationClient(invalidConfigWithYaml).chatCompletion()
+    ).toThrowErrorMatchingInlineSnapshot(`
+     "Prompt Template YAML does not conform to the defined type. Validation errors: [
+       {
+         "code": "invalid_type",
+         "expected": "object",
+         "received": "undefined",
+         "path": [
+           "spec"
+         ],
+         "message": "Required"
+       }
+     ]"
+    `);
   });
 
   it('calls chatCompletion with grounding configuration', async () => {