+++ packages/orchestration/src/orchestration-client.ts
@@ -1,6 +1,8 @@
 import { executeRequest } from '@sap-ai-sdk/core';
 import { resolveDeploymentId } from '@sap-ai-sdk/ai-api/internal.js';
 import { createLogger } from '@sap-cloud-sdk/util';
+import yaml from 'yaml';
+import { promptTemplatePostRequestSchema } from '@sap-ai-sdk/prompt-registry';
 import { OrchestrationStream } from './orchestration-stream.js';
 import { OrchestrationStreamResponse } from './orchestration-stream-response.js';
 import { OrchestrationResponse } from './orchestration-response.js';
@@ -8,6 +10,7 @@ import {
   constructCompletionPostRequest,
   constructCompletionPostRequestFromJsonModuleConfig
 } from './util/index.js';
+import type { TemplatingChatMessage } from './client/api/schema/index.js';
 import type {
   HttpResponse,
   CustomRequestConfig
@@ -42,12 +45,13 @@ export class OrchestrationClient {
     private deploymentConfig?: ResourceGroupConfig,
     private destination?: HttpDestinationOrFetchOptions
   ) {
-    try {
-      if (typeof config === 'string') {
-        JSON.parse(config);
-      }
-    } catch (error) {
-      throw new Error(`Could not parse JSON: ${error}`);
+    if (typeof config === 'string') {
+      this.validateJsonConfig(config);
+    } else {
+      this.config =
+        typeof config.templating === 'string'
+          ? this.parseAndMergeTemplating(config) // parse and assign if templating is a string
+          : config;
     }
   }
 
@@ -159,5 +163,54 @@ export class OrchestrationClient {
       ._pipe(OrchestrationStream._processTokenUsage, response);
 
     return response;
+  }
+
+  /**
+   * Validate if a string is valid JSON.
+   * @param config - The JSON string to validate.
+   */
+  private validateJsonConfig(config: string): void {
+    try {
+      JSON.parse(config);
+    } catch (error) {
+      throw new Error(`Could not parse JSON: ${error}`);
+    }
+  }
+
+  /**
+   * Parse and merge templating into the config object.
+   * @param config - The orchestration module configuration with templating either as object or string.
+   * @returns The updated and merged orchestration module configuration.
+   * @throws Error if the YAML parsing fails or if the parsed object does not conform to the expected schema.
+   */
+  private parseAndMergeTemplating(
+    config: OrchestrationModuleConfig
+  ): OrchestrationModuleConfig {
+    let parsedObject;
+    if (typeof config.templating === 'string' && !config.templating.trim()) {
+      throw new Error('Templating YAML string must be non-empty.');
+    }
+    try {
+      parsedObject = yaml.parse(config.templating as string);
+    } catch (error) {
+      throw new Error(`Error parsing YAML: ${error}`);
+    }
+
+    const result = promptTemplatePostRequestSchema.safeParse(parsedObject);
+    if (!result.success) {
+      throw new Error(
+        `Prompt Template YAML does not conform to the defined type. Validation errors: ${result.error}`
+      );
+    }
+    const { template, defaults, response_format, tools } = result.data.spec;
+    return {
+      ...config,
+      templating: {
+        template: template as TemplatingChatMessage,
+        ...(defaults && { defaults }),
+        ...(response_format && { response_format }),
+        ...(tools && { tools })
+      }
+    };
   }
 }