+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ai-sdk/langchain
 
+## 1.12.0
+
+### Minor Changes
+
+- 5360c8c: [Compatibility Note] Use `LangchainOrchestrationModuleConfig` to type the configuration object passed to the LangChain orchestration client.
+
+### Patch Changes
+
+- Updated dependencies [5360c8c]
+  - @sap-ai-sdk/orchestration@1.12.0
+  - @sap-ai-sdk/ai-api@1.12.0
+  - @sap-ai-sdk/core@1.12.0
+  - @sap-ai-sdk/foundation-models@1.12.0
+
 ## 1.11.0
 
 ### Minor Changes